MNT: split out FS independent AD code
